Shuffle around the Poseidon package
