+++ src/com/projectposeidon/README
@@ -1,0 +1,1 @@
+We will no longer store classes under the name of the developer who created it. Current legacy stuff may stay in its current position tho.
