--- src/com/projectposeidon/moderator_man/HTTPResponse.java
@@ -1,23 +1,0 @@
-package com.projectposeidon.moderator_man;
-
-public class HTTPResponse
-{
-    private String response;
-    private int responseCode;
-    
-    public HTTPResponse(String response, int responseCode)
-    {
-        this.response = response;
-        this.responseCode = responseCode;
-    }
-    
-    public String getResponse()
-    {
-        return response;
-    }
-    
-    public int getResponseCode()
-    {
-        return responseCode;
-    }
-}
