--- src/com/projectposeidon/moderator_man/SessionRequestRunnable.java
@@ -1,6 +1,0 @@
-package com.projectposeidon.moderator_man;
-
-public interface SessionRequestRunnable
-{
-    public void callback(int responseCode, String username, String uuid, String ip);
-}
