+++ src/com/projectposeidon/util/HTTPResponse.java
@@ -1,0 +1,23 @@
+package com.projectposeidon.util;
+
+public class HTTPResponse
+{
+    private String response;
+    private int responseCode;
+    
+    public HTTPResponse(String response, int responseCode)
+    {
+        this.response = response;
+        this.responseCode = responseCode;
+    }
+    
+    public String getResponse()
+    {
+        return response;
+    }
+    
+    public int getResponseCode()
+    {
+        return responseCode;
+    }
+}
