+++ src/com/projectposeidon/util/SessionRequestRunnable.java
@@ -1,0 +1,6 @@
+package com.projectposeidon.util;
+
+public interface SessionRequestRunnable
+{
+    public void callback(int responseCode, String username, String uuid, String ip);
+}
