+++ src/net/minecraft/server/NetLoginHandler.java
@@ -9,7 +9,7 @@ import java.net.Socket;
 import java.util.Random;
 import java.util.logging.Logger;
 
-import static com.projectposeidon.marcindevelopment.Release2Beta.deserializeAddress;
+import static com.projectposeidon.util.Release2Beta.deserializeAddress;
 
 public class NetLoginHandler extends NetHandler {
